[pylint] Ensure path and line are present
We need the path and the line to create the message, and the regex for
the pylint linter does not ensure it.

In a pylint report, it is possible to find a line containing only `::`
so the script crashes.

This commit fixes the problem by updating the `PYLINT_LINE_REGEX` to
force the presence of a path and a line.