Print summary and return exit code 1 if problem found
Before this commit, nothing was printed on the output to know if
something happened or not, and there was no difference in the exit code
if lint problems were found or not.

The exit code statis is important when integrating this tool in a CI, to
mark the PR as failed.

This commit resolves this by:

- adding a return value to the `report` method as a list of boolean, one
for each comment to post on github:
- True if the comment was posted
- False if not
- analyzing this result in the `main` function and displaying a
different message if there is some result or not, and if yes, also
displaying the number of failure if any
- exiting the program with the exit code `1` if lint problems were found
(if they were posted on github or not), and `0` if none were found