Print summary and return exit code 1 if problem found
Before this commit, nothing was printed on the output to know if
something happened or not, and there was no difference in the exit code
if lint problems were found or not.

The exit code statis is important when integrating this tool in a CI, to
mark the PR as failed.

This commit resolves this by:

- adding a return value to the `report` method, a tuple with:
  - Reporter name
  - Number of line with problem(s)
  - Number not already present on github
  - Raised overflow, or None
  - Number that cannot be pushed on github
- passing this result from the `main` function to a new `manage_results`
function that will:
  - display a message depending of the result for each reporter
  - return 1 if at least 1 problem was found in the PR (even if already
pushed on Github on a previous call), and 0 if not