Comment on the PR with problems that aren't in the diff.
In SwiftLint there are rules that can be enforced like the length of a file.
When these hit, the error message will be on the last line even though the
commit caused the failure.

This will still comment if it can on the top of the file with any
failing messages.